requirements.txt
fastapi
python-telegram-bot==21.4
openai
uvicorn
